kernel: synchronize with the latest 'scope-minimized manual hooks'
-v1.5
__https://github.com/backslashxx/KernelSU/issues/5
____sucompat: add is_su_allowed
____sucompat: ksu_sucompat_common -> ksu_sucompat_user_common
____sucompat: amend logic within ksu_handle_execveat_sucompat
____sucompat: move ksu_handle_execve_sucompat before ksu_handle_execveat_sucompat
____sucompat: retain some minute logic / style (ksu_sucompat_non_kp)
____sucompat: identify 'sh' (ksu_handle_execveat_sucompat)
____ksud: update comment

-9/22/25: Fix mis-merge (ksu_handle_execveat_sucompat: sh -> ksud_path; drop sh identifier)

<><><>

Description addendums:
kernel: sucompat: increase reliability, commonize and micro-optimize (https://github.com/tiann/KernelSU/pull/2656)
https://github.com/backslashxx/KernelSU/commit/c4530ac9f323f84cb543846d8dd4aebe4acafdc2

...

Stale: https://github.com/tiann/KernelSU/pull/2656

<><><>

kernel: ksud: migrate ksud execution to security_bprm_check (https://github.com/tiann/KernelSU/pull/2653)
https://github.com/backslashxx/KernelSU/commit/cd07ae28259ce06c3d2c8cff33f06f092b47399f

...

as for envp, we pass the pointer then hunt for it when needed

My reasoning on adding a fallback on usercopy is that on some devices a fault
happens, and it copies garbled data. On my creation of this, I actually had to lock
that _nofault copy on a spinlock as a way to mimic preempt_disable/enable without
actually doing it. As per user reports, no failed _nofault copies anyway but we
have-to-have a fallback for resilience.

...

With that, It also provides an inlined copy_from_user_nofault for < 5.8.

...

<><><>

-https://gitlab.com/pershoot/susfs4ksu/-/tree/gki-android14-6.1-dev